[fix] nuxt build で大量の警告が出る
WARN  Though the "loose" option was set to "false" in your @babel/preset-env config
ugin-proposal-private-property-in-object since the "loose" mode option was set to "true" for @babel/plugin-proposal-pri│
vate-methods.                                                                                                          │
The "loose" option must be the same for @babel/plugin-proposal-class-properties, @babel/plugin-proposal-private-methods│
 and @babel/plugin-proposal-private-property-in-object (when they are enabled): you can silence this warning by explici│
tly adding                                                                                                             │
        ["@babel/plugin-proposal-private-property-in-object", { "loose": true }]                                       │
to the "plugins" section of your Babel config.                                                                         │